refactor: better reducers hierarchy for scheduling
