+++ packages/ui/src/ui/store/selectors/scheduling/index.ts
@@ -4,9 +4,9 @@ import {createSelector} from 'reselect';
 
 export const getSchedulingIsFinalLoadingState = createSelector(
     [
-        (store: RootState) => store.scheduling.loading,
-        (store: RootState) => store.scheduling.loaded,
-        (store: RootState) => store.scheduling.error,
+        (store: RootState) => store.scheduling.scheduling.loading,
+        (store: RootState) => store.scheduling.scheduling.loaded,
+        (store: RootState) => store.scheduling.scheduling.error,
     ],
     (loading, loaded, error) => {
         const status = calculateLoadingStatus(loading, loaded, error);
