+++ packages/ui/src/ui/store/selectors/scheduling/monitor.ts
@@ -3,4 +3,4 @@ import {RootState} from '../../reducers';
 type MonitorChartStuts = {[chartType: string]: boolean};
 
 export const getSchedulingMonitorChartStates = (state: RootState): MonitorChartStuts =>
-    state.scheduling.monitorChartStatus;
+    state.scheduling.scheduling.monitorChartStatus;
