+++ packages/ui/src/ui/store/selectors/scheduling/scheduling-operations.ts
@@ -3,18 +3,18 @@ import {createSelector} from 'reselect';
 import {RootState} from '../../../store/reducers';
 
 export const getSchedulingOperations = (state: RootState) =>
-    state.schedulingOperations.rawOperations;
+    state.scheduling.operations.rawOperations;
 export const getSchedulingOperationsTree = (state: RootState) =>
-    state.schedulingOperations.rawOperationsTree;
-export const getSchedulingOperationsError = (state: RootState) => state.schedulingOperations.error;
+    state.scheduling.operations.rawOperationsTree;
+export const getSchedulingOperationsError = (state: RootState) => state.scheduling.operations.error;
 export const getSchedulingOperationsLoading = (state: RootState) =>
-    state.schedulingOperations.loading;
+    state.scheduling.operations.loading;
 export const getSchedulingOperationsLoaded = (state: RootState) =>
-    state.schedulingOperations.loading;
+    state.scheduling.operations.loading;
 export const getSchedulingOperationsExpandedPools = (state: RootState) =>
-    state.schedulingOperations.expandedPools;
+    state.scheduling.operations.expandedPools;
 export const getSchedulingOperationsLoadAll = (state: RootState) =>
-    state.schedulingOperations.loadAllOperations;
+    state.scheduling.operations.loadAllOperations;
 
 export const getSchedulingOperationsLoadingStatus = createSelector(
     [getSchedulingOperationsLoading, getSchedulingOperationsLoaded],
